Move isinstance to QPU.run (and tests)
